fix invalid phpdoc

+++ src/Drivers/EloquentEntitySet.php
@@ -151,9 +151,7 @@ class EloquentEntitySet extends EntitySet implements CountInterface, CreateInter
      */
     public function getTable(): string
     {
-        $model = $this->getModel();
-
-        return $model->getTable();
+        return $this->getModel()->getTable();
     }
 
     /**
@@ -168,7 +166,7 @@ class EloquentEntitySet extends EntitySet implements CountInterface, CreateInter
     /**
      * Read an Eloquent model
      * @param  PropertyValue  $key  Model key
-     * @return Entity|null Entity
+     * @return Entity Entity
      */
     public function read(PropertyValue $key): Entity
     {
@@ -257,6 +255,10 @@ class EloquentEntitySet extends EntitySet implements CountInterface, CreateInter
     public function delete(PropertyValue $key): void
     {
         $model = $this->getModelByKey($key);
+
+        if (null === $model) {
+            throw new NotFoundException('entity_not_found', 'Entity not found');
+        }
 
         try {
             $model->delete();
